+++ src/main/java/ru/vyarus/dropwizard/orient/OrientServerBundle.java
@@ -78,7 +78,7 @@ public class OrientServerBundle<T extends Configuration & HasOrientServerConfigu
             return;
         }
 
-        final EmbeddedOrientServer orientServer = new EmbeddedOrientServer(conf);
+        final EmbeddedOrientServer orientServer = new EmbeddedOrientServer(conf, environment.getObjectMapper());
         environment.lifecycle().manage(orientServer);
         environment.healthChecks().register("orient-server", new OrientServerHealthCheck());
         if (conf.isAdminServlet()) {